+++ www/wp-content/themes/hittheroad-2021/woocommerce/single-product/product-image.php
@@ -12,7 +12,7 @@
  *
  * @see     https://docs.woocommerce.com/document/template-structure/
  * @package WooCommerce\Templates
- * @version 9.7.0
+ * @version 10.5.0
  */
 
 use Automattic\WooCommerce\Enums\ProductType;